requirements.txt
apache-airflow
apache-airflow-providers-mysql
apache-airflow-providers-common-sql
mysql-connector-python
